Add PageContent to Ticket.page.tsx
