--- src/2_FSD_pages/TicketPage/components/PageContent/PageContent.module.scss
@@ -1,4 +1,0 @@
-.PageContent {
-	padding-top: 95px;
-	padding-bottom: 122px;
-}
