+++ src/2_FSD_pages/TicketPage/components/TicketPage/Ticket.page.tsx
@@ -1,14 +1,18 @@
-import { getRouteTicket } from "@config/router"
+import { getRouteChooseTrain, getRoutePassengers, getRouteTicket } from "@config/router"
 import { BreadcrumbsLine } from "@features/BreadcrumbsLine"
 import { useSetCurrentDirectionByUrl } from "@features/FillingFormCurrentDirection"
+import { useGetFormForSearchOfDirectionsDataForRequestSelector } from "@features/FillingFormForSearchOfDirections"
+import { createQueryParams } from "@helpers/createLinkWithParams/createLinkWithParams.helper"
 import { TypedMemo } from "@sharedProviders/TypedMemo"
 import { ContainerLayout } from "@ui/layout"
 import { Page } from "@ui/Page"
+import { DirectionDetails } from "@widgets/DirectionDetails"
 import { Footer } from "@widgets/Footer"
 import { Header } from "@widgets/Header"
-import { useMemo } from "react"
+import { PageContent } from "@widgets/PageContent"
+import { useCallback } from "react"
 import { useParams } from "react-router"
-import { PageContent } from "../PageContent/PageContent"
+import { useNavigate } from "react-router-dom"
 
 const TicketPage = TypedMemo(() => {
 	const { id } = useParams<{ id: string }>()
@@ -16,26 +20,30 @@ const TicketPage = TypedMemo(() => {
 
 	useSetCurrentDirectionByUrl(pagePath.route)
 
-	const pageContent = (
-		<>
+	const navigate = useNavigate()
+	const formParametres = useGetFormForSearchOfDirectionsDataForRequestSelector()
+
+	const onBackHandler = useCallback(() => {
+		navigate(createQueryParams(getRouteChooseTrain().route, formParametres))
+	}, [formParametres, navigate])
+
+	return (
+		<Page>
 			<Header
 				backgroundType={"search"}
 				pagePath={pagePath.route}
 			></Header>
 			<BreadcrumbsLine stage={"tickets"} />
 			<ContainerLayout>
-				<PageContent />
+				<PageContent
+					nextLink={getRoutePassengers().route}
+					onBackCustomHandler={onBackHandler}
+				>
+					<DirectionDetails />
+				</PageContent>
 			</ContainerLayout>
-		</>
-	)
-
-	const footer = useMemo(() => <Footer pagePath={pagePath.route} />, [pagePath.route])
-
-	return (
-		<Page
-			content={pageContent}
-			footer={footer}
-		/>
+			<Footer pagePath={pagePath.route} />
+		</Page>
 	)
 })
